requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0

# Flask Web Framework
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0
python-socketio>=5.9.0

# Video Streaming
pillow>=10.0.0
imutils>=0.5.4

# Optional but Recommended
python-dotenv>=1.0.0
eventlet>=0.33.0  # For better WebSocket performance
gunicorn>=21.2.0  # For production deployment

# System Dependencies (install separately)
# FFmpeg - Required for HLS streaming
# Download from: https://ffmpeg.org/download.html
